Organize Satteri plugins
The Satteri implementations currently live in astro.config.ts, making
the framework configuration hard to scan and plugin ownership unclear.

This commit will move processor behavior into focused modules under
src/plugins/satteri, expose the ordered plugin list from an index, and
update tests to import the plugin API from its new home.

- https://satteri.bruits.org/docs/plugins/
